moved snapshotStyledComponent into separate file because compile errors
