+++ src/components/StationCard/StationCard.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { snapshotStyledComponent } from 'helpers/helpers';
+import { snapshotStyledComponent } from 'helpers/snapshotStyledComponent';
 import StationCard from './StationCard';
 import renderer from 'react-test-renderer';
